fix(ci): resolve Windows build errors and MSVC warnings
This commit fixes a compilation error in PointsContainer.cpp and addresses several MSVC-specific warnings:
- Fixed incorrect variable capture in PointsContainer.cpp lambda.
- Added [[maybe_unused]] to unreferenced parameters in GlobalLookAndFeel, EnvelopeListSection, PluginProcessor, PresetSection, and SettingsSection.
- Added explicit static_cast<float> to Graphics::drawLine and fillEllipse calls to resolve C4244 warnings.
- Fixed sign conversion and shadowing warnings in PluginProcessor and EnvelopeListSection.
- Updated startDragging to the recommended overload to avoid deprecation warning.

+++ src/ui/sections/settings/SettingsPageBase.h
@@ -24,7 +24,7 @@ public:
 
         // Underline Divider
         g.setColour(T_COL(border).withAlpha(0.3f));
-        g.drawLine(20, 45, getWidth() - 20, 45, 1.0f);
+        g.drawLine(20.0f, 45.0f, static_cast<float>(getWidth()) - 20.0f, 45.0f, 1.0f);
     }
 
     /** Helper to get a standard content area for settings controls */
